+++ src/modules/filesystem/filesystem.h
@@ -23,7 +23,7 @@ const char* lovrFilesystemGetRealDirectory(const char* path);
 const char* lovrFilesystemGetRequirePath();
 const char* lovrFilesystemGetCRequirePath();
 const char* lovrFilesystemGetSaveDirectory(void);
-size_t lovrFilesystemGetSize(const char* path);
+uint64_t lovrFilesystemGetSize(const char* path);
 const char* lovrFilesystemGetSource(void);
 bool lovrFilesystemGetUserDirectory(char* buffer, size_t size);
 bool lovrFilesystemGetWorkingDirectory(char* dest, unsigned int size);
